# [6] graph/db: use batch fetching to improve ForEachNode* performance