--- a/sqldb/sqlc/queries/graph.sql
+++ b/sqldb/sqlc/queries/graph.sql
@@ -446,6 +446,59 @@ WHERE version = $1
 ORDER BY scid DESC
 LIMIT 1;
 
+-- name: ListChannelsForNodeIDs :many
+SELECT sqlc.embed(c),
+       n1.pub_key AS node1_pubkey,
+       n2.pub_key AS node2_pubkey,
+
+       -- Policy 1
+       -- TODO(elle): use sqlc.embed to embed policy structs
+       --  once this issue is resolved:
+       --  https://github.com/sqlc-dev/sqlc/issues/2997
+       cp1.id AS policy1_id,
+       cp1.node_id AS policy1_node_id,
+       cp1.version AS policy1_version,
+       cp1.timelock AS policy1_timelock,
+       cp1.fee_ppm AS policy1_fee_ppm,
+       cp1.base_fee_msat AS policy1_base_fee_msat,
+       cp1.min_htlc_msat AS policy1_min_htlc_msat,
+       cp1.max_htlc_msat AS policy1_max_htlc_msat,
+       cp1.last_update AS policy1_last_update,
+       cp1.disabled AS policy1_disabled,
+       cp1.inbound_base_fee_msat AS policy1_inbound_base_fee_msat,
+       cp1.inbound_fee_rate_milli_msat AS policy1_inbound_fee_rate_milli_msat,
+       cp1.message_flags AS policy1_message_flags,
+       cp1.channel_flags AS policy1_channel_flags,
+       cp1.signature AS policy1_signature,
+
+       -- Policy 2
+       cp2.id AS policy2_id,
+       cp2.node_id AS policy2_node_id,
+       cp2.version AS policy2_version,
+       cp2.timelock AS policy2_timelock,
+       cp2.fee_ppm AS policy2_fee_ppm,
+       cp2.base_fee_msat AS policy2_base_fee_msat,
+       cp2.min_htlc_msat AS policy2_min_htlc_msat,
+       cp2.max_htlc_msat AS policy2_max_htlc_msat,
+       cp2.last_update AS policy2_last_update,
+       cp2.disabled AS policy2_disabled,
+       cp2.inbound_base_fee_msat AS policy2_inbound_base_fee_msat,
+       cp2.inbound_fee_rate_milli_msat AS policy2_inbound_fee_rate_milli_msat,
+       cp2.message_flags AS policy2_message_flags,
+       cp2.channel_flags AS policy2_channel_flags,
+       cp2.signature AS policy2_signature
+
+FROM graph_channels c
+         JOIN graph_nodes n1 ON c.node_id_1 = n1.id
+         JOIN graph_nodes n2 ON c.node_id_2 = n2.id
+         LEFT JOIN graph_channel_policies cp1
+                   ON cp1.channel_id = c.id AND cp1.node_id = c.node_id_1 AND cp1.version = c.version
+         LEFT JOIN graph_channel_policies cp2
+                   ON cp2.channel_id = c.id AND cp2.node_id = c.node_id_2 AND cp2.version = c.version
+WHERE c.version = $1
+  AND (c.node_id_1 IN (sqlc.slice('node1_ids')/*SLICE:node1_ids*/)
+   OR c.node_id_2 IN (sqlc.slice('node2_ids')/*SLICE:node2_ids*/));
+
 -- name: ListChannelsByNodeID :many
 SELECT sqlc.embed(c),
     n1.pub_key AS node1_pubkey,